+++ Js2IL.Tests/Import/ExecutionTests.cs
@@ -1,0 +1,18 @@
+using Xunit;
+
+namespace Js2IL.Tests.Import
+{
+    public class ExecutionTests : ExecutionTestsBase
+    {
+        public ExecutionTests() : base("Import")
+        {
+        }
+
+        [Fact]
+        public Task Import_BasicImport()
+        {
+            var testName = nameof(Import_BasicImport);
+            return ExecutionTest(testName);
+        }
+    }
+}